more integration in teacher_interface.py
