+++ .streamlit/config.toml
@@ -1,0 +1,7 @@
+[theme]
+base="light"
+primaryColor="#007EA7"
+backgroundColor="#FFFFFF"
+secondaryBackgroundColor="#F0F8FF"
+textColor="#003459"
+font="sans serif"
